backend/requirements.txt

# Core FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Mobile automation
appium-python-client==3.1.0
selenium==4.15.2

# AI and ML
openai==1.3.0
qdrant-client==1.6.0

# Image processing and OCR
Pillow==10.1.0
pytesseract==0.3.10
opencv-python==4.8.1.78

# Monitoring and metrics
prometheus-client==0.19.0

# Security and encryption
cryptography==41.0.7
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Database and storage
sqlalchemy==2.0.23
alembic==1.12.1

# Redis for caching (optional)
redis==5.0.1

# Utilities
numpy==1.24.3
pytest==7.4.3
python-dotenv==1.0.0
httpx==0.25.2
websockets==12.0

# Logging and configuration
structlog==23.2.0
python-json-logger==2.0.7

# Development tools
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Additional utilities
jinja2==3.1.2
markdown2==2.4.10
python-docx==1.1.0
python-pptx==0.6.23
